monolith: code refactor
