+++ projects/monolith/abis.json
@@ -1,7 +1,8 @@
 {
     "lens": {
         "poolsInfo": "function poolsInfo() view returns (tuple(address id, string symbol, bool stable, address token0Address, address token1Address, address gaugeAddress, address bribeAddress, address[] bribeTokensAddresses, address fees, uint256 totalSupply, address feeDistAddress)[])",
-        "poolsReservesInfo": "function poolsReservesInfo(address[] pools) view returns (tuple(address id, address token0Address, address token1Address, uint256 token0Reserve, uint256 token1Reserve, uint8 token0Decimals, uint8 token1Decimals)[])"
+        "poolsReservesInfo": "function poolsReservesInfo(address[] pools) view returns (tuple(address id, address token0Address, address token1Address, uint256 token0Reserve, uint256 token1Reserve, uint8 token0Decimals, uint8 token1Decimals)[])",
+        "veTokensIdsOf": "function veTokensIdsOf(address owner) view returns (uint256[] tokens)"
     },
     "gauge": {
         "balanceOf": "function balanceOf(address owner) view returns (uint256 balance)"
